Convert Rest to TS. Fixes #146
